Better organization to project structure
